Added a list of phrases, converted into tree form.
I used DansGuardian's phraselist. It was written in this awful, incomprehensible
format, so I wrote a ruby script to convert the list to JSON. All of that is
included in a tarball, which I will delete before the next commit.

phraselist-pretty contains the same data as phraselist-ugly, but it has things
like newlines, and indentation. Cabal will build with phraselist-ugly. The
library will more than likely use phraselist-ugly, as well.

phraselist-pretty is included for development purposes.